requirements.txt
anyio==3.6.2
async-timeout==4.0.2
bitcoin-rpc @ git+https://github.com/GrayFinance/python-bitcoin-rpc.git
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
fastapi==0.87.0
h11==0.14.0
htlc @ git+https://github.com/GrayFinance/htlc.git
idna==3.4
importlib-metadata==5.0.0
lnd @ git+https://github.com/GrayFinance/lnd.git
packaging==21.3
pydantic==1.10.2
pyparsing==3.0.9
python-bitcoinlib==0.11.2
python-dotenv==0.21.0
pyzmq==24.0.1
redis==4.3.4
requests==2.28.1
sniffio==1.3.0
starlette==0.21.0
tinydb==4.7.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
wrapt==1.14.1
zipp==3.10.0
